refactor : handleClickChatRoomItem 내부 구현 각 페이지 이동. 관심사의 분리
