+++ renderer/src/components/unit/chat/chatRoomList/ChatList.container.tsx
@@ -1,5 +1,4 @@
 import { CREATE_OPEN_CHAT_PAGE } from '@constants/paths';
-import { useEnterOpenChatRoom } from '@contexts/RoomContext';
 import { RoomType } from '@type/room.types';
 import { useRouter } from 'next/router';
 import ChatListUI from './ChatList.presenter';
@@ -8,29 +7,19 @@ type ChatListContainerPropsType = {
   roomList: RoomType[] | null;
   title: string;
   subTitle: string;
+  onClickChatRoomItem: (room: RoomType) => () => void;
 };
 
 export default function ChatListContainer({
   roomList,
   title,
   subTitle,
+  onClickChatRoomItem,
 }: ChatListContainerPropsType) {
   const router = useRouter();
-  const enterOpenChatRoom = useEnterOpenChatRoom();
 
   const handleClickCreateOpenChat = () => {
     router.push(CREATE_OPEN_CHAT_PAGE);
-  };
-
-  const handleClickRoomItem = (room: RoomType) => () => {
-    if (room.type.includes('open')) {
-      enterOpenChatRoom(room.uid);
-    }
-    router.push(
-      `/chat/${room.uid}?type=${room.type}&title=${
-        room.title || room.users?.[0].nickName
-      }`,
-    );
   };
 
   return (
@@ -38,7 +27,7 @@ export default function ChatListContainer({
       roomList={roomList}
       title={title}
       subTitle={subTitle}
-      onClickRoomItem={handleClickRoomItem}
+      onClickRoomItem={onClickChatRoomItem}
       onClickCreateOpenChat={handleClickCreateOpenChat}
     />
   );
